+++ src/Craftliltplugin.php
@@ -15,16 +15,21 @@ namespace lilthq\craftliltplugin;
 
 use Craft;
 use craft\base\Plugin;
+use craft\controllers\EntriesController;
 use craft\helpers\UrlHelper;
 use LiltConnectorSDK\Api\JobsApi;
 use LiltConnectorSDK\Api\SettingsApi;
 use LiltConnectorSDK\Api\TranslationsApi;
 use LiltConnectorSDK\Configuration;
 use lilthq\craftliltplugin\assets\CraftLiltPluginAsset;
+use lilthq\craftliltplugin\assets\EditEntryAsset;
+use lilthq\craftliltplugin\models\TranslationModel;
 use lilthq\craftliltplugin\parameters\CraftliltpluginParameters;
 use lilthq\craftliltplugin\records\SettingRecord;
 use lilthq\craftliltplugin\services\appliers\ElementTranslatableContentApplier;
 use lilthq\craftliltplugin\services\appliers\field\FieldContentApplier;
+use lilthq\craftliltplugin\services\handlers\CopySourceTextHandler;
+use lilthq\craftliltplugin\services\handlers\CreateDraftHandler;
 use lilthq\craftliltplugin\services\handlers\CreateJobHandler;
 use lilthq\craftliltplugin\services\handlers\CreateTranslationsHandler;
 use lilthq\craftliltplugin\services\handlers\EditJobHandler;
@@ -39,7 +44,7 @@ use lilthq\craftliltplugin\services\mappers\LanguageMapper;
 use lilthq\craftliltplugin\services\providers\ConnectorConfigurationProvider;
 use lilthq\craftliltplugin\services\providers\ElementTranslatableContentProvider;
 use lilthq\craftliltplugin\services\providers\field\FieldContentProvider;
-use lilthq\craftliltplugin\services\repositories\external\ConnectorJobFileRepository;
+use lilthq\craftliltplugin\services\repositories\external\ConnectorFileRepository;
 use lilthq\craftliltplugin\services\repositories\external\ConnectorJobRepository;
 use lilthq\craftliltplugin\services\repositories\external\ConnectorTranslationRepository;
 use lilthq\craftliltplugin\services\repositories\I18NRepository;
@@ -47,7 +52,11 @@ use lilthq\craftliltplugin\services\repositories\JobLogsRepository;
 use lilthq\craftliltplugin\services\repositories\JobRepository;
 use lilthq\craftliltplugin\services\repositories\TranslationRepository;
 use lilthq\craftliltplugin\services\ServiceInitializer;
+use yii\base\ActionEvent;
+use yii\base\Controller;
+use yii\base\Event;
 use yii\base\InvalidConfigException;
+use yii\log\Logger;
 use yii\web\Response;
 
 /**
@@ -66,7 +75,7 @@ use yii\web\Response;
  *
  * @property ConnectorJobRepository $connectorJobRepository
  * @property ConnectorTranslationRepository $connectorTranslationRepository
- * @property ConnectorJobFileRepository $connectorJobsFileRepository
+ * @property ConnectorFileRepository $connectorJobsFileRepository
  * @property JobRepository $jobRepository
  * @property TranslationRepository $translationRepository
  * @property ConnectorConfigurationProvider $connectorConfigurationProvider
@@ -91,6 +100,8 @@ use yii\web\Response;
  * @property TranslationFailedHandler $translationFailedHandler
  * @property CreateTranslationsHandler $createTranslationsHandler
  * @property RefreshJobStatusHandler $refreshJobStatusHandler
+ * @property CreateDraftHandler $createDraftHandler
+ * @property CopySourceTextHandler $copySourceTextHandler
  * @property ServiceInitializer $serviceInitializer
  */
 class Craftliltplugin extends Plugin
@@ -210,7 +221,9 @@ class Craftliltplugin extends Plugin
             $this->connectorKey = $connectorKey;
         }
 
-        Craft::$app->getView()->registerAssetBundle(CraftLiltPluginAsset::class);
+        if (Craft::$app->request->isCpRequest) {
+            Craft::$app->getView()->registerAssetBundle(CraftLiltPluginAsset::class);
+        }
 
         $this->setComponents([
             'serviceInitializer' => ServiceInitializer::class
@@ -224,6 +237,48 @@ class Craftliltplugin extends Plugin
                 ['name' => $this->name]
             ),
             __METHOD__
+        );
+
+        Event::on(
+            EntriesController::class,
+            Controller::EVENT_BEFORE_ACTION,
+            static function (ActionEvent $event) {
+                if ($event->action->id !== 'edit-entry') {
+                    return $event;
+                }
+
+                $params = Craft::$app->request->resolve();
+                if (!isset($params[1]['entryId'])) {
+                    return $event;
+                }
+
+                $entryId = (int)$params[1]['entryId'];
+
+                $translations = Craftliltplugin::getInstance()->translationRepository->findInProgressByElementId(
+                    $entryId
+                );
+
+                if (!empty($translations)) {
+                    Craft::$app->getView()->registerAssetBundle(EditEntryAsset::class);
+
+                    $jobIds = array_map(static function (TranslationModel $translation) {
+                        return $translation->jobId;
+                    }, $translations);
+
+                    Craft::$app->view->registerJs(
+                        sprintf(
+                            'new CraftliltPlugin.EntryEditWarning(%s);',
+                            json_encode([
+                                'translationInProgress' => true,
+                                'entry' => $entryId,
+                                'jobs' => array_unique($jobIds)
+                            ], 4194304)
+                        )
+                    );
+                }
+
+                return $event;
+            }
         );
     }