+++ src/elements/Job.php
@@ -72,6 +72,14 @@ class Job extends Element
 
     public function beforeDelete(): bool
     {
+        $translationRecords = TranslationRecord::findAll(['jobId' => $this->id]);
+
+        array_map(static function (TranslationRecord $t) {
+            Craft::$app->elements->deleteElementById(
+                $t->translatedDraftId
+            );
+        }, $translationRecords);
+
         JobRecord::deleteAll(['id' => $this->id]);
         return parent::beforeDelete();
     }
@@ -186,14 +194,21 @@ class Job extends Element
     public function isInstantFlow(): bool
     {
         return strtolower($this->translationWorkflow) === strtolower(
-            SettingsResponse::LILT_TRANSLATION_WORKFLOW_INSTANT
+            CraftliltpluginParameters::TRANSLATION_WORKFLOW_INSTANT
         );
     }
 
     public function isVerifiedFlow(): bool
     {
         return strtolower($this->translationWorkflow) === strtolower(
-            SettingsResponse::LILT_TRANSLATION_WORKFLOW_VERIFIED
+            CraftliltpluginParameters::TRANSLATION_WORKFLOW_VERIFIED
+        );
+    }
+
+    public function isCopySourceTextFlow(): bool
+    {
+        return strtolower($this->translationWorkflow) === strtolower(
+            CraftliltpluginParameters::TRANSLATION_WORKFLOW_COPY_SOURCE_TEXT
         );
     }
 
@@ -457,9 +472,11 @@ class Job extends Element
         return CraftliltpluginParameters::JOB_EDIT_PATH . '/' . $this->id;
     }
 
-    /**
-     * @return TranslationModel[]
-     */
+    public function getTranslationWorkflowLabel(): string
+    {
+        return Craft::t('craft-lilt-plugin', strtolower($this->translationWorkflow));
+    }
+
     public function getTranslations(): array
     {
         if (!empty($this->_elements)) {
@@ -523,8 +540,6 @@ class Job extends Element
         $actions[] = [
             'type' => Delete::class,
         ];
-
-        #$actions[] = LiltJobSetStatus::class;
 
         return $actions;
     }