+++ src/parameters/CraftliltpluginParameters.php
@@ -9,6 +9,8 @@ declare(strict_types=1);
 
 namespace lilthq\craftliltplugin\parameters;
 
+use Craft;
+use LiltConnectorSDK\Model\SettingsResponse;
 use lilthq\craftliltplugin\services\listeners\AfterDraftAppliedListener;
 use lilthq\craftliltplugin\services\listeners\RegisterDefaultTableAttributesListener;
 use lilthq\craftliltplugin\services\listeners\RegisterElementTypesListener;
@@ -39,20 +41,20 @@ class CraftliltpluginParameters
     public const SETTINGS_TABLE_NAME = '{{%lilt_settings}}';
     public const JOB_LOGS_TABLE_NAME = '{{%lilt_jobs_logs}}';
 
-    public const CRAFT_FIELDS_MATRIX            = 'craft\fields\Matrix';
-    public const CRAFT_FIELDS_PLAINTEXT         = 'craft\fields\PlainText';
-    public const CRAFT_FIELDS_LIGHTSWITCH       = 'craft\fields\Lightswitch';
-    public const CRAFT_REDACTOR_FIELD           = 'craft\redactor\Field';
-    public const CRAFT_FIELDS_TABLE             = 'craft\fields\Table';
-    public const CRAFT_FIELDS_SUPER_TABLE       = 'verbb\supertable\fields\SuperTableField';
-    public const CRAFT_FIELDS_RADIOBUTTONS      = 'craft\fields\RadioButtons';
-    public const CRAFT_FIELDS_DROPDOWN          = 'craft\fields\Dropdown';
-    public const CRAFT_FIELDS_MULTISELECT       = 'craft\fields\MultiSelect';
-    public const CRAFT_FIELDS_CHECKBOXES        = 'craft\fields\Checkboxes';
-    public const CRAFT_FIELDS_BASEOPTIONSFIELD  = 'craft\fields\BaseOptionsField';
-    public const BENF_NEO_FIELD                 = 'benf\neo\Field';
-    public const LINKIT_FIELD                   = 'fruitstudios\linkit\fields\LinkitField';
-    public const COLOUR_SWATCHES_FIELD          = 'percipioglobal\colourswatches\fields\ColourSwatches';
+    public const CRAFT_FIELDS_MATRIX = 'craft\fields\Matrix';
+    public const CRAFT_FIELDS_PLAINTEXT = 'craft\fields\PlainText';
+    public const CRAFT_FIELDS_LIGHTSWITCH = 'craft\fields\Lightswitch';
+    public const CRAFT_REDACTOR_FIELD = 'craft\redactor\Field';
+    public const CRAFT_FIELDS_TABLE = 'craft\fields\Table';
+    public const CRAFT_FIELDS_SUPER_TABLE = 'verbb\supertable\fields\SuperTableField';
+    public const CRAFT_FIELDS_RADIOBUTTONS = 'craft\fields\RadioButtons';
+    public const CRAFT_FIELDS_DROPDOWN = 'craft\fields\Dropdown';
+    public const CRAFT_FIELDS_MULTISELECT = 'craft\fields\MultiSelect';
+    public const CRAFT_FIELDS_CHECKBOXES = 'craft\fields\Checkboxes';
+    public const CRAFT_FIELDS_BASEOPTIONSFIELD = 'craft\fields\BaseOptionsField';
+    public const BENF_NEO_FIELD = 'benf\neo\Field';
+    public const LINKIT_FIELD = 'fruitstudios\linkit\fields\LinkitField';
+    public const COLOUR_SWATCHES_FIELD = 'percipioglobal\colourswatches\fields\ColourSwatches';
 
     public const LISTENERS = [
         AfterDraftAppliedListener::class,
@@ -61,4 +63,30 @@ class CraftliltpluginParameters
         RegisterDefaultTableAttributesListener::class,
         RegisterTableAttributesListener::class,
     ];
+
+    public const TRANSLATION_WORKFLOW_INSTANT = SettingsResponse::LILT_TRANSLATION_WORKFLOW_INSTANT;
+    public const TRANSLATION_WORKFLOW_VERIFIED = SettingsResponse::LILT_TRANSLATION_WORKFLOW_VERIFIED;
+    public const TRANSLATION_WORKFLOW_COPY_SOURCE_TEXT = 'COPY_SOURCE_TEXT';
+
+
+    public static function getTranslationWorkflows(): array
+    {
+        return [
+            strtolower(self::TRANSLATION_WORKFLOW_INSTANT) =>
+                Craft::t(
+                    'craft-lilt-plugin',
+                    strtolower(self::TRANSLATION_WORKFLOW_INSTANT)
+                ),
+            strtolower(self::TRANSLATION_WORKFLOW_VERIFIED) =>
+                Craft::t(
+                    'craft-lilt-plugin',
+                    strtolower(self::TRANSLATION_WORKFLOW_VERIFIED)
+                ),
+            strtolower(self::TRANSLATION_WORKFLOW_COPY_SOURCE_TEXT) =>
+                Craft::t(
+                    'craft-lilt-plugin',
+                    strtolower(self::TRANSLATION_WORKFLOW_COPY_SOURCE_TEXT)
+                )
+        ];
+    }
 }