+++ src/services/handlers/CreateTranslationsHandler.php
@@ -4,28 +4,39 @@ declare(strict_types=1);
 
 namespace lilthq\craftliltplugin\services\handlers;
 
+use craft\base\ElementInterface;
 use lilthq\craftliltplugin\elements\Job;
 use lilthq\craftliltplugin\records\TranslationRecord;
 
 class CreateTranslationsHandler
 {
+    /**
+     * @param Job $job
+     * @param array $sourceContent
+     * @param int $elementId
+     * @param int $versionId
+     * @param ElementInterface[] $drafts
+     * @return bool
+     */
     public function __invoke(
         Job $job,
-        array $sourceContent,
+        array $sourceContents,
         int $elementId,
-        int $versionId
+        int $versionId,
+        array $drafts
     ): bool {
         $translationRecords = array_values(
             array_map(
-                static function (int $targetSiteId) use ($job, $sourceContent, $elementId, $versionId) {
+                static function (int $targetSiteId) use ($job, $sourceContents, $elementId, $versionId, $drafts) {
                     return new TranslationRecord([
                         'jobId' => $job->id,
                         'elementId' => $elementId,
                         'versionId' => $versionId,
                         'sourceSiteId' => $job->sourceSiteId,
                         'targetSiteId' => $targetSiteId,
-                        'sourceContent' => $sourceContent,
+                        'sourceContent' => $sourceContents[$targetSiteId],
                         'status' => TranslationRecord::STATUS_IN_PROGRESS,
+                        'translatedDraftId' => $drafts[$targetSiteId]->getId()
                     ]);
                 },
                 $job->getTargetSiteIds()