+++ tests/integration/controllers/job/GetJobEditFormControllerCest.php
@@ -164,6 +164,7 @@ class GetJobEditFormControllerCest
                 'translationWorkflowsOptions' => [
                     'instant' => 'Instant',
                     'verified' => 'Verified',
+                    'copy_source_text' => 'Copy source text'
                 ],
                 'availableSites' => [
                     0 => [
@@ -205,7 +206,7 @@ class GetJobEditFormControllerCest
                     'elementIds' => sprintf('["%d"]', $job->getElementIds()[0]),
                     'versions' => '[]',
                     'dueDate' => null,
-                    'translationWorkflow' => 'INSTANT',
+                    'translationWorkflow' => 'instant',
                     'tempId' => null,
                     'draftId' => null,
                     'revisionId' => null,