+++ CHANGELOG.md
@@ -4,6 +4,37 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.0 - 2022-04-19
+## 0.4.3 - 2022-08-17
+### Fixed
+- Removed value updating for BaseOption & Lightswitch fields
+
+## 0.4.2 - 2022-08-17
+### Fixed
+- Translation applier wasn't able to find source element for site id  
+- Wrong language formatting when locale is empty
+
+## 0.4.1 - 2022-08-16
 ### Added
-- Initial release
+- Error logging for all exception catches  
+
+## 0.4.0 - 2022-08-13
+### Added
+- Remove entry translation drafts on job removal
+- Empty content now excluded from translation source body
+- Warning for job entries, when translation in progress (on job creation) 
+- Fixed: Empty I18N entry creation
+- Fixed: Translation preview issue (error while encoding of json)
+- Fixed: Hide warning when element index not exist
+- Fixed: Argument 2 passed to craft\services\Drafts::createDraft() must be of the type int, null given
+
+## 0.3.1 - 2022-08-13
+### Added
+- Updated connector sdk dependency 
+
+## 0.3.0 - 2022-08-03
+### Added
+- Fixed applying translations for nested fields 
+
+## 0.2.1 - 2022-08-03
+### Added
+- Initial beta release